+++ src/scripts/main.js
@@ -1,4 +1,3 @@
-import logInForm from "./chats/lr_Users.js";
-import userAPI from "./chats/userData.js";
+import logInForm from "./users/users.js";
 
 logInForm();
