+++ src/scripts/users/users.js
@@ -1,14 +1,15 @@
 import addNewUserEvent from "./usersFactory.js";
+import logInUserEvent from "./userLogInFactory.js";
 
 const outputEl = document.querySelector(".targetThis");
 
 const logInForm = () => {
-  document
-    .querySelector(".logIn__Register")
-    .addEventListener("click", (event) => {
-      console.log("button clicked");
-      if (event.target.id === "logIn") {
-        outputEl.innerHTML = `
+	document
+		.querySelector(".logIn__Register")
+		.addEventListener("click", (event) => {
+			console.log("button clicked");
+			if (event.target.id === "logIn") {
+				outputEl.innerHTML = `
             <div class="logIn__form">
                 <h3>Log in to Nutshell!</h3>
                 <form action="" id="">
@@ -24,8 +25,8 @@ const logInForm = () => {
                 <button class="button__logIn" id="button__logIn">Log In</button>
             </div>
         `;
-      } else {
-        outputEl.innerHTML = `
+			} else {
+				outputEl.innerHTML = `
         <div class="register__Form">
             <h3>Treat yo'self to Nutshell!</h3>
             <form action="" id="">
@@ -45,8 +46,9 @@ const logInForm = () => {
             <button class="button__register" id="button__register">Register</button>
         </div>
     `;
-      }
-      addNewUserEvent();
-    });
+			}
+			logInUserEvent();
+			// addNewUserEvent();
+		});
 };
 export default logInForm;
